+++ activation-tool/TridevyatoeActivationTool/App.xaml.cs
@@ -1,0 +1,5 @@
+using System.Windows;
+
+namespace TridevyatoeActivationTool;
+
+public partial class App : Application { }